# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWC_test doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/resources/{id}/bookings": {
            "get": {
                "tags": [
                    "Index"
                ],
                "summary": "Get all bookings for a specific resource",
                "description": "Endpoint to retrieve all bookings associated with a specific resource ID.",
                "operationId": "75962eca4593c818622a10fa99a6493f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource whose bookings are to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "The ID of the booking.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "resource_id": {
                                                        "description": "The ID of the resource.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "description": "The ID of the user.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "start_time": {
                                                        "description": "Start time of the booking.",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-01 12:00:00"
                                                    },
                                                    "end_time": {
                                                        "description": "End time of the booking.",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-01 14:00:00"
                                                    },
                                                    "created_at": {
                                                        "description": "Timestamp when the booking was created.",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-01 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "description": "Timestamp when the booking was last updated.",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-01 12:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}